fixed: only check availability for videodb items (don't try to delete "Add Source" ;))
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@18481 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIMediaWindow.cpp
@@ -827,7 +827,7 @@ bool CGUIMediaWindow::OnClick(int iItem)
     else
     {
       const CFileItemPtr pItem = m_vecItems->Get(iItem);
-      if (!XFILE::CFile::Exists(pItem->GetVideoInfoTag()->m_strFileNameAndPath))
+      if (pItem->IsVideoDb() && !XFILE::CFile::Exists(pItem->GetVideoInfoTag()->m_strFileNameAndPath))
       {
         if (!CGUIWindowVideoNav::DeleteItem(pItem.get(),true))
           return true;
